umount, eject, and remove

+++ thread.h
@@ -1,10 +1,6 @@
 #ifndef HEADER_THREAD
 #define HEADER_THREAD
 
-#include <linux/delay.h>
-#include <linux/kthread.h>
-#include <asm/errno.h>
-
 extern int nas_timer_ticks;
 
 #define TIMER_INTERVAL 10 // sec
@@ -14,16 +10,16 @@ extern int nas_timer_ticks;
 /*
  * Interuptible sleep in kthread
  */
-#define kthread_ssleep(sec)                               \
-	if (1) {                                              \
-		set_current_state(TASK_INTERRUPTIBLE);            \
-		if (kthread_should_stop()) {                      \
-			set_current_state(TASK_RUNNING);              \
-			break;                                        \
-		}                                                 \
-		schedule_timeout(msecs_to_jiffies(sec*1000));     \
-		set_current_state(TASK_RUNNING);                  \
-	} else                                                \
+#define kthread_ssleep(sec)                                      \
+	if (1) {                                                     \
+		set_current_state(TASK_INTERRUPTIBLE);                   \
+		if (kthread_should_stop()) {                             \
+			set_current_state(TASK_RUNNING);                     \
+			break;                                               \
+		}                                                        \
+		schedule_timeout(msecs_to_jiffies((int)(sec*1000)));     \
+		set_current_state(TASK_RUNNING);                         \
+	} else                                                       \
 		do {} while (0)
 
 int start_nas_mon(void);
